% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtune.mJAM.R
\name{xtune_mJAM}
\alias{xtune_mJAM}
\title{Regularized regression incorporating external information for summary statistics analysis}
\usage{
xtune_mJAM(
  betas.Gx,
  N.Gx,
  Geno,
  eaf.Gy = NULL,
  Z = as.matrix(rep(1, ncol(Geno[[1]]))),
  c = 0.5,
  sigma.square = NULL,
  alpha.est.init = rep(0, ncol(Z) + 1),
  maxstep = 100,
  tolerance = 0.001,
  maxstep_inner = 50,
  tolerance_inner = 0.1,
  compute.likelihood = FALSE,
  verbosity = FALSE,
  standardize = TRUE,
  intercept = TRUE,
  epsilon = 3
)
}
\arguments{
\item{betas.Gx}{A list of numeric vectors that contains the summary statistics from each ethnic population}

\item{N.Gx}{A list of numeric that contains sample sizes in all original GWAS studies}

\item{Geno}{A list of matrices with individual-level SNP dosage data in each study/population as reference data. Each column corresponds to a SNP. Note that the order of the study/population should be matched in all lists.}

\item{eaf.Gy}{The effect allele frequency of the SNPs in betas.Gx}

\item{Z}{Prior information matrix about the predictors/SNPs (\eqn{p} rows, each corresponding to a SNP in X; \eqn{q} columns of external information about the predictors, such as prior biological importance). If Z is the grouping of predictors, it is best if user codes it as a dummy variable (i.e. each column indicating whether predictors belong to a specific group).}

\item{c}{The elastic-net mixing parameter ranging from 0 to 1. When  \eqn{c} = 1, the model corresponds to Lasso. When \eqn{c} is set to 0, it corresponds to Ridge. For values between 0 and 1 (with a default of 0.5), the model corresponds to the elastic net.}

\item{sigma.square}{A user-supplied noise variance estimate. Typically, this is left unspecified, and the function automatically computes an estimated sigma square values using R package \code{selectiveinference}.}
}
\value{

}
\description{
Regularized regression incorporating external information for summary statistics analysis
}
